% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdapipeline.R
\name{fdapipeline}
\alias{fdapipeline}
\alias{<fdapipeline>-package}
\title{fdapipeline}
\description{
Contains functions for the analysis of circadian data.
    Use four spaces when indenting paragraphs within the Description.
}
\author{
Sarah Lock \email{sarah.lock@york.ac.uk}
}
