% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata_example.R
\docType{data}
\name{mydata_example}
\alias{mydata_example}
\title{Arabidopsis thaliana luciferase experiment}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{link to data?}
}
\usage{
data(mydata_example)
}
\description{
Data from a luciferase experiment looking at response to changes in photo period.
Arabidopsis, with data on phyB mutants and WS-2 parent lines, the outcome is raw counts per second (cps) of each plants luminscence at increments over 166 hours. Plants were either subjected to entrainment under short day (8 hours light, 16 hours dark) or long day (16 hours light, 8 hours dark) then shifted to the opposite condition during the expermient. Controls were also carried out. #'
}
\examples{
data(mydata_example)
}
\references{
paper 2024
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{link to paper})
}
\keyword{datasets}
