% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_fun.R
\name{smooth_fun}
\alias{smooth_fun}
\title{Compute and filter circadian time series data}
\usage{
smooth_fun(x, time, outlier = TRUE, shape = TRUE)
}
\arguments{
\item{x}{Matrix containing data values, each row is a time point and each column corresponds to a sample.}

\item{time}{matrix containing the time values.}

\item{outlier}{A positive integer specifying the derivative the curves will be evaluated at for outlier detection.The default is the first derivative.}

\item{shape}{A logical indicating whether shape outlier detection should be performed}
}
\value{
A smooth functional data object
}
\description{
The function is designed to smooth multiple curves represented by time-series data. It normalises each curve and organizes them into a list, removing any missing values. It then creates basis functions and functional parameters for each curve using B-splines the default is set to the number of timepoints and sets a roughness penalty. Next, it generates smooth basis functions for each curve individually. After defining a new time vector covering the range of all curves, it evaluates the smoothed curves under this new time vector. Subsequently, it combines all evaluated curves and performs an outlier detection process using the Total Variation (TV) smooth method from \code{fdaoutliers}. Any shape outliers are removed, and the remaining curves are smoothed again collectively. Finally, the function returns the smoothed curves along with the new time vector and the total curves matrix.
}
\details{
This function allows estimation of functions of time series data using cubic b-spline basis system.
}
\examples{
data(mydata_example)
data <- mydata_example[,-1]
time <- as.matrix(mydata_example[,1])
smooth_curves <- smooth_fun(data, time, shape = FALSE, outlier = 1)



}
