% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTraits.R
\docType{package}
\name{TimeTraits-package}
\alias{TimeTraits}
\alias{TimeTraits-package}
\title{TimeTraits: Functional Data Analysis Pipeline, Extracting Functional Traits from Biological Time-Series Data.}
\description{
Provides tools for analysing circadian time-series using functional
data analysis, including smoothing, functional principal component
analysis, and extraction of group-level traits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/scllock/TimeTraits}
  \item Report bugs at \url{https://github.com/scllock/TimeTraits/issues}
}

}
\author{
\strong{Maintainer}: Sarah Lock \email{sarah.lock@york.ac.uk}

}
