% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeTraits.R
\name{TimeTraits}
\alias{TimeTraits}
\alias{<TimeTraits>-package}
\title{TimeTraits}
\description{
Contains functions for the analysis of biological time-series data.
    Use four spaces when indenting paragraphs within the Description.
}
\author{
Sarah Lock \email{sarah.lock@york.ac.uk}
}
