% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_curves.R
\name{filter_curves}
\alias{filter_curves}
\title{Filter curves based on criteria}
\usage{
filter_curves(
  x,
  time,
  from = min(time, na.rm = TRUE),
  to = max(time, na.rm = TRUE),
  min_lum = 0,
  periodogram_length = NULL
)
}
\arguments{
\item{x}{Matrix containing data values, each row is a time point and each column corresponds to a sample, can be a sparse matrix.}

\item{time}{Vector containing the time values.}

\item{from}{An integer specifying the time point to begin analysis. The default is the minimum of the time vector.}

\item{to}{An integer specifying the time point to end analysis. The default is the maximum of the time vector.}

\item{min_lum}{An integer specifying the minimum luminescence value to exclude samples. Default is 0.}

\item{periodogram_length}{An integer specifying the number of hours for the Lomb-Scargle periodogram to be performed. Default is NULL and set to the length of \code{time}.}
}
\value{
A matrix array is returned with filtered columns and rows based on selected parameters.
install.packages("devtools") and install.packages(roxygen2)
}
\description{
The filter_curves function begins determining the time range for analysis and subsetting the data accordingly (default is full length of input). Samples with luminescence values below a specified threshold are filtered out (default is 0). The function applies linear regression to estimate and then remove trends from the curves, focusing on oscillatory components. Lomb-Scargle periodogram analysis is performed to detect rhythmic patterns in the detrended data (default is full length of input), followed by peak detection to identify significant periodic components. Samples failing to meet a significance threshold p > 0.001 in the periodogram analysis are filtered out. Finally, the function constructs and returns a filtered data matrix containing the specified time vector and curves that pass the filtering criteria, ready for further analysis or visualisation. Overall, \code{filter_curves} provides a comprehensive tool for preprocessing and analysing luminescence data, ensuring that only relevant and significant samples are retained for downstream analyses.
}
\details{
This function allows for the filtering of luminescence data based on time points and a minimum luminescence threshold in order for the further analysis. Using lomb-scargle periodogram to determine if each sample contains rhythmic data the non-rhythmic data is removed. See X function
}
\examples{
data(mydata_example)
data <- mydata_example[,-1]
time <- as.matrix(mydata_example[,1])
o <- filter_curves(data, time, from = 24, to = 144, min_lum = 200, periodogram_length = 48)
}
